Add simple integration test and test project
Made a simple play project using the recommended play file
hierarchy and configured it to build using the plugin.

The project has a controller which is refered to by a single
route and renders a single template. This lets us test for successful
template and route compilation against a java class.

SmokeTest installs the sample project and checks for errors.